+++ ui/appui-react/src/appui-react/editors/CustomNumberEditor.tsx
