+++ ui/appui-react/src/appui-react/editors/NumericEditor.tsx
