+++ ui/appui-react/src/appui-react/editors/TextEditor.tsx
